Rename cse.py to subexpr_cache.py
'CSE' is compiler-optimization jargon and not widely understood outside
that niche. 'subexpression cache' describes what the module supports in
terms that don't require CS expertise.
