+++ cvxpy/reductions/dnlp2smooth/dnlp2smooth.py
@@ -19,9 +19,9 @@ from cvxpy import problems
 from cvxpy.expressions.expression import Expression
 from cvxpy.problems.objective import Minimize
 from cvxpy.reductions.canonicalization import Canonicalization
-from cvxpy.reductions.cse import UncacheableError, expr_key
 from cvxpy.reductions.dnlp2smooth.canonicalizers import SMOOTH_CANON_METHODS as smooth_canon_methods
 from cvxpy.reductions.inverse_data import InverseData
+from cvxpy.reductions.subexpr_cache import UncacheableError, expr_key
 
 
 class Dnlp2Smooth(Canonicalization):
